pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["amelie"]

[project.urls]
Homepage = "https://www.inter-actief.utwente.nl/"
Repository = "https://github.com/Inter-Actief/amelie"
Issues = "https://github.com/Inter-Actief/amelie/issues"

[project]
name = "amelie"
description = "Le fabuleux destin d'Amélie - De website van Inter-/Actief/. Vernoemd naar de film Amélie."
dynamic = ["version"]
readme = "README.md"
license = "BSD-3-Clause"
license-files = ["LICENSE.md"]

authors = [
    { name = "WWW committee of I.C.T.S.V. Inter-Actief", email = "www@inter-actief.net" }
]

requires-python = ">= 3.10"

dependencies = [
    "Django>=5.2.8,<5.3",
    "importlib-metadata>=1.4.0,<10.0",
    "celery>=5.6.0,<5.7",
    "flower>=2.0.1,<2.1",
    "django-celery-results>=2.6,<2.7",

    "django-compressor>=4.1,<4.7",
    "django-localflavor>=5.0,<5.1",
    "django-extensions>=3.2.1,<4.2",
    "django-formtools>=2.4,<2.6",
    "django-auditlog>=3.4.1,<3.5",
    "django-oauth-toolkit>=1.7.1,<3.3",
    "django-cors-headers>=3.13.0,<4.10",
    "django-colorfield",
    "django-environ>=0.13.0,<0.14",

    # Django Modern RPC -- For our JSONRPC API
    "django-modern-rpc>=2.0.0,<2.1",

    # WYSIWYG Editor for Django
    "django-ckeditor>=6.3.2",

    # FCM (Firebase Cloud Messaging)
    # 1.x has breaking changes (old firebase api removed) # TODO: Breaking
    "fcm-django>=0.3.11,<1",

    # GraphQL API
    "graphene-django",
    "django-graphql-jwt",
    "django-filter>=25.2,<26",
    "graphene-django-extras",

    # Used for stripping Markdown notations and indexing the IA Summary website.
    "beautifulsoup4>=4.11.1,<4.15",

    # Sanitizing Markdown strings in amelie template tags
    "nh3>=0.3.3,<0.4",

    # Other random dependencies, to be documented, see issue #556
    "httplib2>=0.20.4,<0.32",
    "icalendar>=7.0.3,<7.1",
    "Markdown>=3.10.2,<3.11",
    "mysqlclient>=2.2.8",

    # Pillow requires libjpeg-dev zlib1g-dev (on Debian)
    "Pillow>=12.1.1,<12.2",
    "python-gitlab>=8.1.0,<8.2",
    "python-dateutil>=2.9.0,<2.10",

    # python-ldap requires libldap-dev libsasl2-dev (on Debian) to build
    "python-ldap>=3.4.3,<3.5",
    "raven>=6.10.0,<6.11",
    "reportlab>=4.4.10",
    "requests>=2.32,<2.33",
    "html2text>=2020.1.16,<2025.4.16",

    # Daphne - to run the website in ASGI mode on a proper webserver
    "daphne>=3.0,<4.3",

    # UWSGI - to run the website in WSGI mode on a proper webserver
    "uwsgi>=2.0,<2.1",

    # QR code generator for personal tab POS (requires Pillow)
    "qrcode>=7.3.1,<8.3",

    # PDF and IPP libraries to upload prints to the website and send them to the IA printer
    "pypdf>=6.1.1,<7",
    "pyipp>=0.17.2,<0.18",

    # Google API Client
    "google-api-python-client>=2.192.0,<3",
    "google-auth>=2.49.1,<2.50",
    "pyOpenSSL>=26.0.0,<26.1.0",

    # Development
    "django-debug-toolbar>=3.7,<6.3",
    "django-rosetta>=0.9.8,<0.11",
    "docutils>=0.19,<0.23",
    "pyinotify",
    "model_bakery>=1.23.3,<1.24",
    "django-silk>=5.5.0,<5.6",

    # Single Sign On - OIDC Client
    "mozilla-django-oidc>=5.0.2,<5.1",

    # Sentry error logging
    "sentry-sdk>=2.54.0,<2.56",

    # Health checks for Django
    "django-health-check[psutil,celery,rabbitmq,redis]>=4.0,<4.3",

    # Allows for CAPTCHA support in forms
    "django-simple-captcha",

    # Verification of uploaded SSH public keys for active members
    "sshpubkeys",

    # Library for managing parts of the IA Matrix server via Claudia
    "mautrix>=0.21.0,<1.0"
]
